requirements.txt
python-dotenv
langchain
PyPDF2
faiss-cpu
streamlit
sentence-transformers
numpy
scikit-learn
torch
-e .